docs: add document for pr29
The documents, docs, how we solve these authentication problems coming from the COVAX side. It can be used as a reference for future content creation. Also, the commit-pr markdown file explained how to the best practice, one of the best practices of structured commit messages and PR messages so that as a developer I don't need to think; it's just a formula for me to follow and apply.